requirements.txt
async-timeout==4.0.2
cachelib==0.9.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cydets==1.1
Deprecated==1.2.13
Flask==2.2.2
Flask-Session==0.4.0
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.3
packaging==21.3
pandas==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
python-dotenv==1.0.0
requests==2.28.1
six==1.16.0
spotipy==2.20.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
